Move ConsumerTools.py to Core/Utilities according to suggestion of F.S.
Also, move unit tests and change the import path in ConsumerModule
